check if file exits

+++ utilities.py
@@ -31,10 +31,11 @@ def predict_all_images():
         tifs = glob.glob("data/evaluation/RGB/*.tif")
         print("{} images found for prediction".format(len(tifs)))
         for tif in tifs:
-                print(tif)
-                df = model.predict_image(path = tif, return_plot=False)
-                if df is None:
-                        continue
-                #save boxes
                 file_path = os.path.splitext(tif)[0] + ".csv"
-                df.to_csv(file_path)
+                if not os.path.exists(file_path):
+                  print(tif)
+                  df = model.predict_image(path = tif, return_plot=False)
+                  if df is None:
+                          continue
+                  #save boxes
+                  df.to_csv(file_path)
